Frida Kahlo Retrospective
Overview

This project is slightly different than others you have encountered thus far on Codecademy. Instead of a step-by-step tutorial, this project contains a series of open-ended requirements which describe the project you’ll be building.

There are many possible ways to correctly fulfill all of these requirements, and you should expect to use the internet, Codecademy, and other resources when you encounter a problem that you cannot easily solve.

Project Goals

Congratulations! You’ve been hired to work on a retrospective of Frida Kahlo’s work at a major museum in Mexico.

Your job is to put together the audio tour, but in order to do that, you need to create a list of each painting featured in the exhibit, the date it was painted, and its spot on the tour.

Use your knowledge of Python lists to create a master list of each painting, its date, and its audio tour ID.