Python Lists: Medical Insurance Estimation Project
In this project, you will examine how factors such as age, sex, BMI, number of children, and smoking status contribute to medical insurance costs.

Note: while insurance companies do use BMI in their calculations, and that is reflected in this project, BMI is not necessarily an accurate predictor of health. As data scientists, we should always be skeptical of quantitative measures like BMI that reduce complex phenomena to a single number.

You will apply your new knowledge of Python Lists to store insurance cost data in a list as well as compare estimated insurance costs to actual insurance costs.